refactor: Article, BlockController, EditableBlock 컴포넌트 로직 분리
